# M2-Challenge-Fantaye-Aynalem/openapi (1).yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /subtract:
    post:
      tags:
        - math-solution-controller
      operationId: subtract
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
  /multiply:
    post:
      tags:
        - math-solution-controller
      operationId: multiply
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
  /divide:
    post:
      tags:
        - math-solution-controller
      operationId: divide
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
  /add:
    post:
      tags:
        - math-solution-controller
      operationId: add
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
  /randomMonth:
    get:
      tags:
        - random-month-controller
      operationId: getRandomMonth
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RandomMonth'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
  /month/{monthNumber}:
    get:
      tags:
        - month-converter-controller
      operationId: getMonth
      parameters:
        - name: monthNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
        status:
          type: integer
          format: int32
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
    MathSolution:
      required:
        - operand1
        - operand2
      type: object
      properties:
        operand1:
          type: integer
          format: int32
        operand2:
          minimum: 1
          type: integer
          format: int32
        answer:
          type: integer
          format: int32
        operation:
          type: string
    RandomMonth:
      required:
        - name
        - number
      type: object
      properties:
        number:
          type: integer
          format: int32
        name:
          type: string